revert: keep zenoh subscribe_all conflating; spy stays protocol-agnostic
Reverses the ratified subscribe_all non-conflation spec change (Ivan's call):
- zenoh subscribe_all: restore original latest-per-topic drain over dimos/**
- spec: drop SubscribeLatestMixin/subscribe_latest entirely
- rerun bridge: back to subscribe_all (its original conflating consumer)
- spy is transport-agnostic — it just calls bus.subscribe_all(), accepting
  whatever delivery semantics each transport provides (LCM: every message;
  zenoh: latest-per-topic). No raw/session-level special-casing.
- drop the now-invalid contract tests: subscribe_all-delivers-every[zenoh]
  and subscribe_latest[*]; docstrings updated to match.

Keeps the three Greptile fixes already on feat/ivan/spy (default_sources
degradation, web-mode filter forwarding, lcmspy-web routing).
