+++ dimos/visualization/rerun/test_viewer_integration.py
@@ -126,9 +126,6 @@ class ExplicitPubSubOverride:
     def subscribe_all(self, callback):
         return lambda: None
 
-    def subscribe_latest(self, callback):
-        return lambda: None
-
 
 class TestBridgePubsubResolution:
     def test_legacy_lcm_pubsubs_defers_to_transport_default(self):
